requirements.txt
vosk
sounddevice
scipy
faiss-cpu
numpy
langdetect
coqui-tts
pyaudio
sentence-transformers
pydub